히트맵 오류 수정완료

+++ src/main/java/com/smoking_map/smoking_map/web/dto/admin/HeatmapDto.java
@@ -2,15 +2,14 @@ package com.smoking_map.smoking_map.web.dto.admin;
 
 import lombok.Getter;
 import java.math.BigDecimal;
-import java.io.Serializable;
 
 @Getter
 public class HeatmapDto {
-    private final BigDecimal latitude;
-    private final BigDecimal longitude;
+    private final double latitude;
+    private final double longitude;
 
     public HeatmapDto(BigDecimal latitude, BigDecimal longitude) {
-        this.latitude = latitude;
-        this.longitude = longitude;
+        this.latitude = latitude.doubleValue();
+        this.longitude = longitude.doubleValue();
     }
 }
